An infinite loop happens when drawing wavy underline for small text at a very large x-position https://bugs.webkit.org/show_bug.cgi?id=286831 rdar://141022987
Reviewed by Simon Fraser.

If the text is displayed at x which is larger than 2^23 and the wavy step is less
than 0.5, the for-loop in strokeWavyTextDecoration() will not terminate. We will
keep adding PathSegements to the Path till WebKit jetsams.

The fix is to use `double` for for-loop control variable so it can add the small
step and we eventually exit the for-loop.

* LayoutTests/fast/text/text-large-x-small-size-wavy-style-expected.txt: Added.
* LayoutTests/fast/text/text-large-x-small-size-wavy-style.html: Added.
* Source/WebCore/rendering/TextDecorationPainter.cpp:
(WebCore::strokeWavyTextDecoration):

Canonical link: https://commits.webkit.org/289659@main

+++ Source/WebCore/rendering/TextDecorationPainter.cpp
@@ -99,7 +99,7 @@ static void strokeWavyTextDecoration(GraphicsContext& context, const FloatRect& 
     FloatPoint controlPoint1(0, yAxis + wavyStrokeParameters.controlPointDistance);
     FloatPoint controlPoint2(0, yAxis - wavyStrokeParameters.controlPointDistance);
 
-    for (float x = x1; x + 2 * wavyStrokeParameters.step <= x2;) {
+    for (double x = x1; x + 2 * wavyStrokeParameters.step <= x2;) {
         controlPoint1.setX(x + wavyStrokeParameters.step);
         controlPoint2.setX(x + wavyStrokeParameters.step);
         x += 2 * wavyStrokeParameters.step;
